+++ CHANGELOG.md
@@ -21,8 +21,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://bitbucket.org/atlassian/aws-resources/branches/compare/master%0Drelease-1.10.2
+[Unreleased]: https://bitbucket.org/atlassian/aws-resources/branches/compare/master%0Drelease-1.11.0
 
+## [1.11.0] - 2023-02-01
+[1.11.0]: https://bitbucket.org/atlassian/aws-resources/branches/compare/release-1.11.0%0Drelease-1.10.2
 ### Added
 - Add logging ssh access in `AwaitingEc2`. [JPERF-954].
 - Grant default permissions to EC2 in `SshAmiMod`. Resolve [JPERF-932].